タスクの簡易選択機能を、モーダルを開かずに選択するように変更。 DashboardContainerに実装を移す部分まで対応。
